test: switch alert tests to react testing library
